@@ -9,8 +9,8 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>21</maven.compiler.source>
-        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.source>23</maven.compiler.source>
+        <maven.compiler.target>23</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <stanford.nlp.version>4.5.8</stanford.nlp.version>
         <jackson.version>2.18.2</jackson.version>
@@ -186,8 +186,28 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.11.0</version>
                 <configuration>
-                    <source>21</source>
-                    <target>21</target>
+                    <source>23</source>
+                    <target>23</target>
+                    <showWarnings>false</showWarnings>
+                    <quiet>true</quiet>
+                </configuration>
+            </plugin>
+
+            <!-- Configure test output -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>3.2.5</version>
+                <configuration>
+                    <redirectTestOutputToFile>true</redirectTestOutputToFile>
+                    <trimStackTrace>true</trimStackTrace>
+                    <reportsDirectory>${project.build.directory}/surefire-reports</reportsDirectory>
+                    <consoleOutputReporter>
+                        <disable>true</disable>
+                    </consoleOutputReporter>
+                    <statelessTestsetReporter implementation="org.apache.maven.plugin.surefire.extensions.junit5.JUnit5Xml30StatelessReporter">
+                        <disable>true</disable>
+                    </statelessTestsetReporter>
                 </configuration>
             </plugin>
 
@@ -247,7 +267,6 @@
                         </configuration>
                     </execution>
                     <!-- LevelDB Browser JAR -->
-                    <!--
                     <execution>
                         <id>browser-jar</id>
                         <phase>package</phase>
@@ -263,7 +282,6 @@
                             </transformers>
                         </configuration>
                     </execution>
-                    -->
                     <!-- Index Runner JAR -->
                     <!--
                     <execution>
@@ -310,23 +328,6 @@
                     <mainClass>com.example.QueryCLI</mainClass>
                     <vmArgs>-Xss16m -XX:MaxRAMPercentage=70</vmArgs>
                 </configuration>
-            </plugin>
-
-            <!-- Surefire plugin -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>3.2.2</version>
-                <configuration>
-                    <argLine>-javaagent:${settings.localRepository}/org/mockito/mockito-core/${mockito.version}/mockito-core-${mockito.version}.jar</argLine>
-                </configuration>
-                <dependencies>
-                    <dependency>
-                        <groupId>org.mockito</groupId>
-                        <artifactId>mockito-core</artifactId>
-                        <version>${mockito.version}</version>
-                    </dependency>
-                </dependencies>
             </plugin>
 
             <!-- ANTLR4 Plugin -->
