@@ -1,6 +1,12 @@
 package com.example.query;
 
 import com.example.query.model.*;
+import com.example.query.model.condition.Contains;
+import com.example.query.model.condition.Dependency;
+import com.example.query.model.condition.Logical;
+import com.example.query.model.condition.Ner;
+import com.example.query.model.condition.Temporal;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
@@ -8,6 +14,7 @@ import org.junit.jupiter.api.DisplayName;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -24,15 +31,17 @@ class QuerySemanticValidatorTest {
     @Test
     @DisplayName("Valid query should pass validation")
     void validQueryShouldPassValidation() throws QueryParseException {
-        Query query = new Query("wikipedia", 
+        Query query = new Query(
+            "wikipedia", 
             Arrays.asList(
-                new ContainsCondition("test"),
-                new NerCondition("PERSON", "Einstein", false)
+                new Contains("test"),
+                Ner.of("PERSON")
             ),
             Collections.singletonList(new OrderSpec("date", OrderSpec.Direction.DESC)),
             Optional.of(10),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
         assertDoesNotThrow(() -> validator.validate(query));
@@ -41,12 +50,14 @@ class QuerySemanticValidatorTest {
     @Test
     @DisplayName("Empty CONTAINS value should throw exception")
     void emptyContainsValueShouldThrow() {
-        Query query = new Query("wikipedia",
-            Collections.singletonList(new ContainsCondition("")),
+        Query query = new Query(
+            "wikipedia",
+            Collections.singletonList(new Contains("")),
             Collections.emptyList(),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
         assertThrows(QueryParseException.class, () -> validator.validate(query));
@@ -55,12 +66,14 @@ class QuerySemanticValidatorTest {
     @Test
     @DisplayName("Invalid NER type should throw exception")
     void invalidNerTypeShouldThrow() {
-        Query query = new Query("wikipedia",
-            Collections.singletonList(new NerCondition("INVALID_TYPE", "test", false)),
+        Query query = new Query(
+            "wikipedia",
+            Collections.singletonList(Ner.of("INVALID_TYPE")),
             Collections.emptyList(),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
         QueryParseException e = assertThrows(QueryParseException.class, 
@@ -71,15 +84,17 @@ class QuerySemanticValidatorTest {
     @Test
     @DisplayName("Variable type mismatch should throw exception")
     void variableTypeMismatchShouldThrow() {
-        Query query = new Query("wikipedia",
+        Query query = new Query(
+            "wikipedia",
             Arrays.asList(
-                new NerCondition("PERSON", "scientist", true),
-                new NerCondition("ORGANIZATION", "scientist", true)
+                Ner.withVariable("PERSON", "?scientist"),
+                Ner.withVariable("ORGANIZATION", "?scientist")
             ),
             Collections.emptyList(),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
         QueryParseException e = assertThrows(QueryParseException.class, 
@@ -93,166 +108,137 @@ class QuerySemanticValidatorTest {
         LocalDateTime start = LocalDateTime.now();
         LocalDateTime end = start.minusDays(1); // End before start
 
-        Query query = new Query("wikipedia",
-            Collections.singletonList(new TemporalCondition(start, end)),
+        Query query = new Query(
+            "wikipedia",
+            Collections.singletonList(new Temporal(start, end)),
             Collections.emptyList(),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
-        QueryParseException e = assertThrows(QueryParseException.class, 
+        QueryParseException e = assertThrows(QueryParseException.class,
             () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("Start date must be before end date"));
+        assertTrue(e.getMessage().contains("End date must be after start date"));
     }
 
     @Test
     @DisplayName("Empty dependency components should throw exception")
     void emptyDependencyComponentsShouldThrow() {
-        Query query = new Query("wikipedia",
-            Collections.singletonList(new DependencyCondition("", "nsubj", "cat")),
+        Query query = new Query(
+            "wikipedia",
+            Collections.singletonList(new Dependency("", "nsubj", "object")),
             Collections.emptyList(),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
-        QueryParseException e = assertThrows(QueryParseException.class, 
+        QueryParseException e = assertThrows(QueryParseException.class,
             () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("cannot have empty components"));
+        assertTrue(e.getMessage().contains("Empty dependency component"));
     }
 
     @Test
     @DisplayName("Empty order by field should throw exception")
     void emptyOrderByFieldShouldThrow() {
-        Query query = new Query("wikipedia",
+        Query query = new Query(
+            "wikipedia",
             Collections.emptyList(),
-            Collections.singletonList(new OrderSpec("")),
+            Collections.singletonList(new OrderSpec("", OrderSpec.Direction.ASC)),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
-        QueryParseException e = assertThrows(QueryParseException.class, 
+        QueryParseException e = assertThrows(QueryParseException.class,
             () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("Order by field cannot be empty"));
+        assertTrue(e.getMessage().contains("Empty order by field"));
     }
 
     @Test
     @DisplayName("Invalid limit should throw exception")
     void invalidLimitShouldThrow() {
-        Query query = new Query("wikipedia",
+        Query query = new Query(
+            "wikipedia",
             Collections.emptyList(),
             Collections.emptyList(),
-            Optional.of(0),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Optional.of(-1),
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
-        QueryParseException e = assertThrows(QueryParseException.class, 
+        QueryParseException e = assertThrows(QueryParseException.class,
             () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("Limit must be greater than 0"));
+        assertTrue(e.getMessage().contains("Invalid limit"));
     }
-    
+
     @Test
     @DisplayName("Unbound variable in ORDER BY should throw exception")
     void unboundVariableInOrderByShouldThrow() {
-        Query query = new Query("wikipedia",
-            Arrays.asList(
-                new ContainsCondition("test")
-            ),
-            Collections.singletonList(new OrderSpec("?unboundVar", OrderSpec.Direction.ASC)),
+        Query query = new Query(
+            "wikipedia",
+            Collections.singletonList(new Contains("test")),
+            Collections.singletonList(new OrderSpec("?person", OrderSpec.Direction.ASC)),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
-        QueryParseException e = assertThrows(QueryParseException.class, 
+        QueryParseException e = assertThrows(QueryParseException.class,
             () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("Unbound variable in ORDER BY"));
+        assertTrue(e.getMessage().contains("Unbound variable"));
     }
-    
+
     @Test
     @DisplayName("Bound variable in ORDER BY should pass validation")
     void boundVariableInOrderByShouldPass() {
-        Query query = new Query("wikipedia",
-            Arrays.asList(
-                new NerCondition("PERSON", "scientist", true)
-            ),
-            Collections.singletonList(new OrderSpec("?scientist", OrderSpec.Direction.ASC)),
+        Query query = new Query(
+            "wikipedia",
+            Collections.singletonList(Ner.withVariable("PERSON", "?person")),
+            Collections.singletonList(new OrderSpec("?person", OrderSpec.Direction.ASC)),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
         assertDoesNotThrow(() -> validator.validate(query));
     }
-    
+
     @Test
     @DisplayName("Complex nested conditions should correctly track bound variables")
     void complexNestedConditionsShouldTrackBoundVariables() {
-        // Create a logical condition with a NOT condition inside
-        NerCondition nerCondition = new NerCondition("PERSON", "scientist", true);
-        NotCondition notCondition = new NotCondition(new ContainsCondition("irrelevant"));
-        LogicalCondition logicalCondition = new LogicalCondition(
-            LogicalCondition.LogicalOperator.AND,
-            Arrays.asList(nerCondition, notCondition)
-        );
-        
-        Query query = new Query("wikipedia",
-            Collections.singletonList(logicalCondition),
-            Collections.singletonList(new OrderSpec("?scientist", OrderSpec.Direction.ASC)),
+        Query query = new Query(
+            "wikipedia",
+            List.of(
+                new Logical(
+                    Logical.LogicalOperator.AND,
+                    Arrays.asList(
+                        Ner.withVariable("PERSON", "?person"),
+                        new Contains("scientist"),
+                        new Logical(
+                            Logical.LogicalOperator.OR,
+                            Arrays.asList(
+                                new Dependency("?person", "nsubj", "discovered"),
+                                new Dependency("?person", "nsubj", "invented")
+                            )
+                        )
+                    )
+                )
+            ),
+            Collections.singletonList(new OrderSpec("?person", OrderSpec.Direction.ASC)),
             Optional.empty(),
-            Optional.empty(),  // granularity
-            Optional.empty()   // granularitySize
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
         );
 
         assertDoesNotThrow(() -> validator.validate(query));
     }
-    
-    // Note: We can't directly test SnippetNode validation until the Query model is updated
-    // to include select columns. The following test is commented out as a reference for
-    // future implementation.
-    
-    /*
-    @Test
-    @DisplayName("Unbound variable in SNIPPET should throw exception")
-    void unboundVariableInSnippetShouldThrow() {
-        // Create a query with a snippet node that references an unbound variable
-        SnippetNode snippetNode = new SnippetNode("?unboundVar", 1);
-        
-        // Once Query model is updated, we'll need to pass the select columns
-        Query query = new Query("wikipedia",
-            Collections.emptyList(),
-            Collections.emptyList(),
-            Optional.empty(),
-            Optional.empty(),
-            Optional.empty()
-        );
-        
-        QueryParseException e = assertThrows(QueryParseException.class, 
-            () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("Unbound variable in SNIPPET"));
-    }
-    
-    @Test
-    @DisplayName("Excessive window size in SNIPPET should throw exception")
-    void excessiveWindowSizeInSnippetShouldThrow() {
-        // Create a query with a snippet node that has a window size that exceeds the maximum
-        SnippetNode snippetNode = new SnippetNode("?person", 10); // Assuming MAX_SNIPPET_WINDOW_SIZE = 5
-        
-        // Once Query model is updated, we'll need to pass the select columns
-        Query query = new Query("wikipedia",
-            Collections.singletonList(new NerCondition("PERSON", "person", true)),
-            Collections.emptyList(),
-            Optional.empty(),
-            Optional.empty(),
-            Optional.empty()
-        );
-        
-        QueryParseException e = assertThrows(QueryParseException.class, 
-            () -> validator.validate(query));
-        assertTrue(e.getMessage().contains("Snippet window size"));
-    }
-    */
 } 
