@@ -3,10 +3,10 @@ package com.example.query.executor;
 import com.example.core.IndexAccess;
 import com.example.core.Position;
 import com.example.core.PositionList;
-import com.example.query.model.ContainsCondition;
 import com.example.query.model.DocSentenceMatch;
-import com.example.query.model.LogicalCondition;
-import com.example.query.model.LogicalCondition.LogicalOperator;
+import com.example.query.model.condition.Contains;
+import com.example.query.model.condition.Logical;
+import com.example.query.model.condition.Logical.LogicalOperator;
 import com.example.query.model.Query;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -75,11 +75,18 @@ public class SentenceGranularityTest {
         when(mockIndexAccess.get(any())).thenReturn(Optional.of(positionList));
         
         // Create a simple CONTAINS condition
-        ContainsCondition condition = new ContainsCondition(List.of("test"));
+        Contains condition = new Contains(List.of("test"));
         
         // Create a query with document granularity
-        Query query = new Query("test");
-        query.addCondition(condition);
+        Query query = new Query(
+            "test",
+            List.of(condition),
+            List.of(),
+            Optional.empty(),
+            Query.Granularity.DOCUMENT,
+            Optional.empty(),
+            List.of()
+        );
         
         // Execute the query with document granularity
         Set<DocSentenceMatch> results = queryExecutor.execute(query, indexes);
@@ -89,8 +96,8 @@ public class SentenceGranularityTest {
         
         // Check that we have document-level matches
         for (DocSentenceMatch match : results) {
-            assertEquals(-1, match.getSentenceId(), "Should be a document-level match");
-            assertTrue(match.getDocumentId() >= 1 && match.getDocumentId() <= 3, 
+            assertEquals(-1, match.sentenceId(), "Should be a document-level match");
+            assertTrue(match.documentId() >= 1 && match.documentId() <= 3, 
                     "Document ID should be between 1 and 3");
         }
     }
@@ -100,12 +107,17 @@ public class SentenceGranularityTest {
         // Set up mock to return our position list for any term
         when(mockIndexAccess.get(any())).thenReturn(Optional.of(positionList));
         
-        // Create a simple CONTAINS condition
-        ContainsCondition condition = new ContainsCondition(List.of("test"));
-        
         // Create a query with sentence granularity
-        Query query = new Query("test", List.of(condition), List.of(), 
-                Optional.empty(), Optional.of(Query.Granularity.SENTENCE), Optional.empty());
+        Contains condition = new Contains("test");
+        Query query = new Query(
+            "test",
+            List.of(condition),
+            List.of(),
+            Optional.empty(),
+            Query.Granularity.SENTENCE,
+            Optional.empty(),
+            List.of()
+        );
         
         // Execute the query with sentence granularity
         Set<DocSentenceMatch> results = queryExecutor.execute(query, indexes);
@@ -116,13 +128,13 @@ public class SentenceGranularityTest {
         // Check that we have sentence-level matches
         int sentenceCount = 0;
         for (DocSentenceMatch match : results) {
-            assertTrue(match.getSentenceId() >= 0, "Should be a sentence-level match");
+            assertTrue(match.sentenceId() >= 0, "Should be a sentence-level match");
             
             // Count sentences per document
-            if (match.getDocumentId() == 1 && match.getSentenceId() == 1) sentenceCount++;
-            if (match.getDocumentId() == 1 && match.getSentenceId() == 2) sentenceCount++;
-            if (match.getDocumentId() == 2 && match.getSentenceId() == 1) sentenceCount++;
-            if (match.getDocumentId() == 3 && match.getSentenceId() == 2) sentenceCount++;
+            if (match.documentId() == 1 && match.sentenceId() == 1) sentenceCount++;
+            if (match.documentId() == 1 && match.sentenceId() == 2) sentenceCount++;
+            if (match.documentId() == 2 && match.sentenceId() == 1) sentenceCount++;
+            if (match.documentId() == 3 && match.sentenceId() == 2) sentenceCount++;
         }
         
         assertEquals(4, sentenceCount, "Should find all 4 sentences");
@@ -133,7 +145,6 @@ public class SentenceGranularityTest {
         // Create two position lists for different terms
         PositionList positionList1 = new PositionList();
         positionList1.add(new Position(1, 1, 10, 15, LocalDate.now())); // Doc 1, Sent 1
-        positionList1.add(new Position(1, 2, 30, 35, LocalDate.now())); // Doc 1, Sent 2
         positionList1.add(new Position(2, 1, 10, 15, LocalDate.now())); // Doc 2, Sent 1
         
         PositionList positionList2 = new PositionList();
@@ -146,16 +157,19 @@ public class SentenceGranularityTest {
         when(mockIndexAccess.get("term2".getBytes())).thenReturn(Optional.of(positionList2));
         
         // Create two CONTAINS conditions
-        ContainsCondition condition1 = new ContainsCondition(List.of("term1"));
-        ContainsCondition condition2 = new ContainsCondition(List.of("term2"));
+        Contains condition1 = new Contains(List.of("term1"));
+        Contains condition2 = new Contains(List.of("term2"));
         
-        // Create a logical AND condition
-        LogicalCondition andCondition = new LogicalCondition(
-                LogicalOperator.AND, Arrays.asList(condition1, condition2));
-        
-        // Create a query with sentence granularity
-        Query query = new Query("test", List.of(andCondition), List.of(), 
-                Optional.empty(), Optional.of(Query.Granularity.SENTENCE), Optional.empty());
+        // Create a query with sentence granularity and window size 1
+        Query query = new Query(
+            "test_db",
+            List.of(new Logical(Logical.LogicalOperator.AND, condition1, condition2)),
+            List.of(),
+            Optional.empty(),
+            Query.Granularity.SENTENCE,
+            Optional.of(0),
+            List.of()
+        );
         
         // Execute the query
         Set<DocSentenceMatch> results = queryExecutor.execute(query, indexes);
@@ -164,8 +178,8 @@ public class SentenceGranularityTest {
         assertEquals(1, results.size(), "Should find 1 sentence with both terms");
         
         DocSentenceMatch match = results.iterator().next();
-        assertEquals(1, match.getDocumentId(), "Should be document 1");
-        assertEquals(1, match.getSentenceId(), "Should be sentence 1");
+        assertEquals(1, match.documentId(), "Should be document 1");
+        assertEquals(1, match.sentenceId(), "Should be sentence 1");
     }
 
     @Test
@@ -184,16 +198,23 @@ public class SentenceGranularityTest {
         when(mockIndexAccess.get("term2".getBytes())).thenReturn(Optional.of(positionList2));
         
         // Create two CONTAINS conditions
-        ContainsCondition condition1 = new ContainsCondition(List.of("term1"));
-        ContainsCondition condition2 = new ContainsCondition(List.of("term2"));
+        Contains condition1 = new Contains(List.of("term1"));
+        Contains condition2 = new Contains(List.of("term2"));
         
         // Create a logical OR condition
-        LogicalCondition orCondition = new LogicalCondition(
+        Logical orCondition = new Logical(
                 LogicalOperator.OR, Arrays.asList(condition1, condition2));
         
         // Create a query with sentence granularity
-        Query query = new Query("test", List.of(orCondition), List.of(), 
-                Optional.empty(), Optional.of(Query.Granularity.SENTENCE), Optional.empty());
+        Query query = new Query(
+            "test_db",
+            List.of(new Logical(Logical.LogicalOperator.OR, condition1, condition2)),
+            List.of(),
+            Optional.empty(),
+            Query.Granularity.SENTENCE,
+            Optional.empty(),
+            List.of()
+        );
         
         // Execute the query
         Set<DocSentenceMatch> results = queryExecutor.execute(query, indexes);
@@ -208,15 +229,64 @@ public class SentenceGranularityTest {
         boolean foundDoc3Sent1 = false;
         
         for (DocSentenceMatch match : results) {
-            if (match.getDocumentId() == 1 && match.getSentenceId() == 1) foundDoc1Sent1 = true;
-            if (match.getDocumentId() == 1 && match.getSentenceId() == 2) foundDoc1Sent2 = true;
-            if (match.getDocumentId() == 2 && match.getSentenceId() == 1) foundDoc2Sent1 = true;
-            if (match.getDocumentId() == 3 && match.getSentenceId() == 1) foundDoc3Sent1 = true;
+            if (match.documentId() == 1 && match.sentenceId() == 1) foundDoc1Sent1 = true;
+            if (match.documentId() == 1 && match.sentenceId() == 2) foundDoc1Sent2 = true;
+            if (match.documentId() == 2 && match.sentenceId() == 1) foundDoc2Sent1 = true;
+            if (match.documentId() == 3 && match.sentenceId() == 1) foundDoc3Sent1 = true;
         }
         
         assertTrue(foundDoc1Sent1, "Should find Doc 1, Sent 1");
         assertTrue(foundDoc1Sent2, "Should find Doc 1, Sent 2");
         assertTrue(foundDoc2Sent1, "Should find Doc 2, Sent 1");
         assertTrue(foundDoc3Sent1, "Should find Doc 3, Sent 1");
+    }
+
+    @Test
+    public void testSentenceGranularityWithWindowSize() throws Exception {
+        // Create two position lists for different terms
+        PositionList positionList1 = new PositionList();
+        positionList1.add(new Position(1, 1, 10, 15, LocalDate.now())); // Doc 1, Sent 1
+        positionList1.add(new Position(2, 1, 10, 15, LocalDate.now())); // Doc 2, Sent 1
+        
+        PositionList positionList2 = new PositionList();
+        positionList2.add(new Position(1, 2, 20, 25, LocalDate.now())); // Doc 1, Sent 2
+        positionList2.add(new Position(2, 2, 30, 35, LocalDate.now())); // Doc 2, Sent 2
+        
+        // Set up mock to return different position lists for different terms
+        when(mockIndexAccess.get("term1".getBytes())).thenReturn(Optional.of(positionList1));
+        when(mockIndexAccess.get("term2".getBytes())).thenReturn(Optional.of(positionList2));
+        
+        // Create two CONTAINS conditions
+        Contains condition1 = new Contains(List.of("term1"));
+        Contains condition2 = new Contains(List.of("term2"));
+        
+        // Create a query with sentence granularity and window size 1
+        Query query = new Query(
+            "test_db",
+            List.of(new Logical(Logical.LogicalOperator.AND, condition1, condition2)),
+            List.of(),
+            Optional.empty(),
+            Query.Granularity.SENTENCE,
+            Optional.of(1),
+            List.of()
+        );
+        
+        // Execute the query
+        Set<DocSentenceMatch> results = queryExecutor.execute(query, indexes);
+        
+        // Verify results - should find matches in Doc 1 (Sent 1,2) and Doc 2 (Sent 1,2)
+        assertEquals(2, results.size(), "Should find 2 pairs of adjacent sentences");
+        
+        // Check that we have the expected sentence pairs
+        boolean foundDoc1 = false;
+        boolean foundDoc2 = false;
+        
+        for (DocSentenceMatch match : results) {
+            if (match.documentId() == 1) foundDoc1 = true;
+            if (match.documentId() == 2) foundDoc2 = true;
+        }
+        
+        assertTrue(foundDoc1, "Should find matches in Doc 1");
+        assertTrue(foundDoc2, "Should find matches in Doc 2");
     }
 } 
