@@ -46,7 +46,7 @@ public final class POSIndexGenerator extends IndexGenerator<AnnotationEntry> {
     @Override
     protected List<AnnotationEntry> fetchBatch(int offset) throws SQLException {
         List<AnnotationEntry> batch = new ArrayList<>();
-        String query = "SELECT a.document_id, a.sentence_id, a.begin_char, a.end_char, a.lemma, a.pos, d.timestamp " +
+        String query = "SELECT a.annotation_id, a.document_id, a.sentence_id, a.begin_char, a.end_char, a.lemma, a.pos, d.timestamp " +
                       "FROM annotations a " +
                       "JOIN documents d ON a.document_id = d.document_id " +
                       "ORDER BY a.document_id, a.sentence_id, a.begin_char LIMIT ? OFFSET ?";
@@ -58,6 +58,7 @@ public final class POSIndexGenerator extends IndexGenerator<AnnotationEntry> {
             try (ResultSet rs = stmt.executeQuery()) {
                 while (rs.next()) {
                     AnnotationEntry entry = new AnnotationEntry(
+                        rs.getInt("annotation_id"),
                         rs.getInt("document_id"),
                         rs.getInt("sentence_id"),
                         rs.getInt("begin_char"),
