@@ -3,9 +3,10 @@ package com.example.query.executor;
 import com.example.core.IndexAccess;
 import com.example.core.Position;
 import com.example.core.PositionList;
-import com.example.query.model.ContainsCondition;
 import com.example.query.model.DocSentenceMatch;
 import com.example.query.model.Query;
+import com.example.query.model.condition.Contains;
+import java.util.Objects;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -20,92 +21,105 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Executor for ContainsCondition.
- * Searches for terms in unigram, bigram, or trigram indexes based on term length.
+ * Executor for CONTAINS conditions.
+ * Handles n-gram pattern matching and variable binding.
+ * 
+ * @see com.example.query.model.condition.Contains
  */
-public class ContainsConditionExecutor implements ConditionExecutor<ContainsCondition> {
-    private static final Logger logger = LoggerFactory.getLogger(ContainsConditionExecutor.class);
+public final class ContainsExecutor implements ConditionExecutor<Contains> {
+    private static final Logger logger = LoggerFactory.getLogger(ContainsExecutor.class);
     
     private static final String UNIGRAM_INDEX = "unigram";
     private static final String BIGRAM_INDEX = "bigram";
     private static final String TRIGRAM_INDEX = "trigram";
     private static final char DELIMITER = IndexAccess.NGRAM_DELIMITER;
 
+    private final String variableName;
+
+    public ContainsExecutor(String variableName) {
+        this.variableName = variableName;
+    }
+
     @Override
-    public Set<DocSentenceMatch> execute(ContainsCondition condition, Map<String, IndexAccess> indexes,
-                               VariableBindings variableBindings, Query.Granularity granularity)
+    public Set<DocSentenceMatch> execute(Contains condition, Map<String, IndexAccess> indexes,
+                               VariableBindings variableBindings, Query.Granularity granularity,
+                               int granularitySize) 
         throws QueryExecutionException {
         
-        List<String> terms = condition.getTerms();
+        logger.debug("Executing CONTAINS condition with {} terms at {} granularity with size {}", 
+                condition.terms().size(), granularity, granularitySize);
         
-        // Log available indexes for debugging
-        logger.info("Available indexes: {}", indexes.keySet());
+        // Validate required indexes
+        if (!indexes.containsKey(UNIGRAM_INDEX)) {
+            throw new QueryExecutionException(
+                "Missing required unigram index",
+                condition.toString(),
+                QueryExecutionException.ErrorType.MISSING_INDEX
+            );
+        }
         
-        // Validate number of terms (1-3)
+        Set<DocSentenceMatch> matches = new HashSet<>();
+        
+        List<String> terms = condition.terms();
+        if (terms.isEmpty()) {
+            logger.warn("CONTAINS condition has no terms, returning empty result set");
+            return matches;
+        }
+        
+        // Check if there are too many terms
         if (terms.size() > 3) {
             throw new QueryExecutionException(
-                "CONTAINS operator supports at most 3 terms, but got " + terms.size(),
-                condition.toString(),
+                "CONTAINS condition supports at most 3 terms, but got " + terms.size() + " terms",
+                "CONTAINS(" + String.join(", ", terms) + ")",
                 QueryExecutionException.ErrorType.INVALID_CONDITION
             );
         }
         
-        // Log whether this is a variable binding condition
-        if (condition.isVariable()) {
-            logger.debug("Executing CONTAINS condition with variable binding '{}' for terms {} at {} granularity", 
-                         condition.getVariableName(), terms, granularity);
-        } else {
-            logger.debug("Executing CONTAINS condition with {} terms at {} granularity", terms.size(), granularity);
+        // Determine if this is a variable binding
+        boolean isVariable = condition.isVariable();
+        
+        // Get the appropriate ngram index based on the size of the terms
+        IndexAccess index = null;
+        if (terms.size() == 1) {
+            index = indexes.get(UNIGRAM_INDEX);
+        } else if (terms.size() == 2) {
+            index = indexes.get(BIGRAM_INDEX);
+        } else if (terms.size() == 3) {
+            index = indexes.get(TRIGRAM_INDEX);
         }
         
-        // Determine which index to use based on term count
-        String indexName;
-        switch (terms.size()) {
-            case 1:
-                indexName = UNIGRAM_INDEX;
-                break;
-            case 2:
-                indexName = BIGRAM_INDEX;
-                break;
-            case 3:
-                indexName = TRIGRAM_INDEX;
-                break;
-            default:
-                // Should never happen due to validation above
-                throw new IllegalStateException("Unexpected number of terms: " + terms.size());
-        }
-        
-        // Check if the index exists
-        IndexAccess index = indexes.get(indexName);
         if (index == null) {
-            logger.error("Required index not found: {}. Available indexes: {}", indexName, indexes.keySet());
             throw new QueryExecutionException(
-                "Required index not found: " + indexName,
-                condition.toString(),
+                "No index found for " + terms.size() + "-gram terms. Available indexes: " + indexes.keySet(),
+                "CONTAINS(" + String.join(", ", terms) + ")",
                 QueryExecutionException.ErrorType.INDEX_ACCESS_ERROR
             );
         }
         
-        // Log which index we're using
-        logger.info("Using {} index for {} terms", indexName, terms.size());
-        
-        // Handle wildcards and construct search patterns
-        Set<String> searchPatterns = constructSearchPatterns(terms);
-        logger.debug("Constructed {} search patterns for terms {}", searchPatterns.size(), terms);
-        
-        // Execute search for each pattern and combine results
-        Set<DocSentenceMatch> result = new HashSet<>();
-        for (String pattern : searchPatterns) {
-            Set<DocSentenceMatch> patternResults = executePatternSearch(
-                pattern, condition.isVariable(), condition.getVariableName(), 
-                index, variableBindings, granularity
+        try {
+            // Construct search patterns based on the terms
+            Set<String> patterns = constructSearchPatterns(terms);
+            
+            // Execute search for each pattern and union the results
+            for (String pattern : patterns) {
+                Set<DocSentenceMatch> patternMatches = executePatternSearch(
+                    pattern, isVariable, variableName, index, variableBindings, granularity);
+                
+                if (!patternMatches.isEmpty()) {
+                    matches.addAll(patternMatches);
+                }
+            }
+            
+            logger.debug("Found {} matches for terms: {}", matches.size(), terms);
+            return matches;
+        } catch (Exception e) {
+            throw new QueryExecutionException(
+                "Error executing CONTAINS condition: " + e.getMessage(),
+                e,
+                condition.toString(),
+                QueryExecutionException.ErrorType.INTERNAL_ERROR
             );
-            result.addAll(patternResults);
         }
-        
-        logger.debug("CONTAINS condition matched {} {}", result.size(), 
-                granularity == Query.Granularity.DOCUMENT ? "documents" : "sentences");
-        return result;
     }
     
     /**
@@ -233,63 +247,47 @@ public class ContainsConditionExecutor implements ConditionExecutor<ContainsCond
                 
             PositionList positionList = positionsOpt.get();
             Set<DocSentenceMatch> matches = new HashSet<>();
-                
+            
             // Process positions based on granularity
             if (granularity == Query.Granularity.DOCUMENT) {
                 // Document granularity - group by document ID
                 Map<Integer, DocSentenceMatch> docMatches = new HashMap<>();
                 
                 for (Position position : positionList.getPositions()) {
-                    int docId = position.getDocumentId();
-                    
-                    // Get or create the document match
-                    DocSentenceMatch match = docMatches.computeIfAbsent(docId, 
-                            id -> new DocSentenceMatch(id));
-                    
-                // Add the position to the match - use original pattern for display
-                match.addPosition(pattern, position);
+                    addDocumentMatch(position, docMatches, variableName);
                     
                     // Handle variable binding
-                if (isVariable) {
-                    bindVariable(variableName, pattern, position, variableBindings);
+                    if (isVariable) {
+                        bindVariable(variableName, pattern, position, variableBindings);
+                    }
                 }
-            }
                 
-            matches.addAll(docMatches.values());
+                matches.addAll(docMatches.values());
             } else {
                 // Sentence granularity - group by document ID and sentence ID
-                Map<String, DocSentenceMatch> sentMatches = new HashMap<>();
+                Map<SentenceKey, DocSentenceMatch> sentMatches = new HashMap<>();
                 
                 for (Position position : positionList.getPositions()) {
-                    int docId = position.getDocumentId();
-                    int sentId = position.getSentenceId();
-                    String key = docId + ":" + sentId;
-                    
-                    // Get or create the sentence match
-                    DocSentenceMatch match = sentMatches.computeIfAbsent(key, 
-                            k -> new DocSentenceMatch(docId, sentId));
-                    
-                // Add the position to the match - use original pattern for display
-                match.addPosition(pattern, position);
+                    addSentenceMatch(position, sentMatches, variableName);
                     
                     // Handle variable binding
-                if (isVariable) {
-                    bindVariable(variableName, pattern, position, variableBindings);
+                    if (isVariable) {
+                        bindVariable(variableName, pattern, position, variableBindings);
+                    }
                 }
+                
+                matches.addAll(sentMatches.values());
             }
+            
+            logger.debug("Found pattern '{}' in {} {}", normalizedPattern, matches.size(), 
+                        granularity == Query.Granularity.DOCUMENT ? "documents" : "sentences");
                 
-            matches.addAll(sentMatches.values());
-        }
-        
-        logger.debug("Found pattern '{}' in {} {}", normalizedPattern, matches.size(), 
-                    granularity == Query.Granularity.DOCUMENT ? "documents" : "sentences");
-                
-        return matches;
+            return matches;
         } catch (Exception e) {
             throw new QueryExecutionException(
-            "Error executing CONTAINS condition for pattern '" + normalizedPattern + "': " + e.getMessage(),
+                "Error executing CONTAINS condition for pattern '" + normalizedPattern + "': " + e.getMessage(),
                 e,
-            "CONTAINS(" + pattern + ")",
+                "CONTAINS(" + pattern + ")",
                 QueryExecutionException.ErrorType.INTERNAL_ERROR
             );
         }
@@ -307,11 +305,12 @@ public class ContainsConditionExecutor implements ConditionExecutor<ContainsCond
                              VariableBindings variableBindings) {
         int docId = position.getDocumentId();
         int sentenceId = position.getSentenceId();
-        int startPos = position.getBeginPosition();
+        int beginPos = position.getBeginPosition();
+        int endPos = position.getEndPosition();
         
-        // Format: term@sentenceId:startPos
-        // This allows extraction of sentence ID from the binding
-        String valueWithPosition = term + "@" + sentenceId + ":" + startPos;
+        // Format: term@beginPos:endPos
+        // This new format makes it easier to use character positions directly for highlighting
+        String valueWithPosition = term + "@" + beginPos + ":" + endPos;
         
         variableBindings.addBinding(docId, variableName, valueWithPosition);
         logger.debug("Bound variable '{}' to '{}' in document {}", 
